fix offline migrations for tutanota 75
when logging in, the app assumes that the
UserSettingsGroupRoot is cached if the UserGroupInfo
is cached. It is required for partialLogin to work at all, so
we have to redownload it after the migration.

Deleting UserGroupInfos as well makes sure that we
fall back to the sync login to redownload the changed entities

Co-authored-by: mup <34790144+murilopereirame@users.noreply.github.com>