fix CalendarEvents offline storage migration for TutanotaModeV74
We need to delete all CalendarEvents since we changed the format for
storing customIds (CalendarEvents use customIds) in the offline
database. All entities with customIds, that are stored in the
offline database (e.g. CalendarEvent, MailSetEntry), are from now on
stored in the offline database using a **base64Ext** encoded id string.