[android] Fix synchronized contacts not appearing in some contact apps
Certain contact apps require the account to be syncable.
In order to do this, we implement a stubbed sync service
and adapter which does nothing and ask Android not to
call it. Preexisting installations of the app will need
to turn contact synchronization off and on again to
receive this fix.
Closes #6568.

Co-authored-by: paw <paw-hub@users.noreply.github.com>