Fix trying to assign TypeRef's to not-migrated entities from offline db
When we try to set correct TypeRef to all aggregates in a type we
traverse the type model. This type model applies to the current version
of the type, not to the one that is in offline db before the migrations
are run.

We separated the typed methods that read the whole table from untyped
ones to avoid the issue above. It has an additional benefit of not
pretending that it returns the entity in its current version.

fix #7524