Allow admin to decrypt user group key via pubAdminEncGKey
Extract AsymmetricCryptoFacade to encapsulate choice of encryption
algorithm.

Remove unauthenticated asymmetric decryption because it was only used
in a deprecated case and is the same implementation as with
authentication.

Enforce the correct protocol version and that it matches the algorithm
of recipient's key pair.

Use existing method to decrypt via adminEncGKey when changing the admin
flag.

Add cryptoWrapper dependency to make testing easier.

Minimize what re need to expose from PqFacade.

Enforce PqKeyPairs type in key rotation facade.