Fix typing for offline DB migrations
Previously migrations been generic based on the entity and was used as
a return type in `Migration`.

This is incorrect as T would refer to the "current" version of the
entity and migrations must move the type between two versions and it
is common that neither of them is the "current" version as referenced
by `TypeRef`.

It would be more safe to say that we expect a certain migration that
only operates on LET or only on ET but in practice all migrations take
an `any` argument and return it rendering all type guards virtually
useless.

In order to not give false sense of safety we removed the generic
parameter from `Migration` type.