Fix email from notification being stuck in the inbox
We removed movable types semantics from EventQueue because move
optimization was interacting poorly with storing email in offline db
from notification process. Move optimization would skip some of the
delete events entirely if the create event was not processed yet. This
would lead to a situation where we would never delete an email that
was put into offline db from notification process but was moved at
some point before login, as cache processing would only download it at
its final destination.

With our change EventQueue does not consider two entities with the same
element id but different list id to be related to each other so the
delete event for the email in inbox will be processed by the cache
normally.

Move semantics for mails will soon be obsolete anyway with the move to
mail sets.

We did not remove the move semantics in other places in the app yet as
it is still a valuable optimization.

Close #7545

Co-authored-by: paw <paw-hub@users.noreply.github.com>