Fix sync progress not shown as finished
Previous change to sync progress tracking did not take into account
event optimization where some batches might be skipped, either
immediately or retroactively (e.g. after we receive a delete event).

One change is to not start counting progress until we know how many
batches will be actually processed (so not until they are added to the
queue).

Another change is to not remove empty event batches from the queue to
not invalidate the expected number of batches that we process.

Note that the batches can still be optimized away completely at the
moment when they are added. That's why the caller must check the return
value of `add()`.

Co-authored-by: paw <paw-hub@users.noreply.github.com>