Fix "Batch does not have events for" error
`batchMod()` was checking the operation on the wrong update.

Some move semantics optimizations were still left out and were not
valid, they might have been not valid in the past too. Since we do not
consider delete + create a move operation anymore we removed the
optimization. Since `_lastOperationForEntity` was changed it was writing
the wrong batch for an id in a specific scenario.

Fix #7552

Co-authored-by: paw <paw-hub@users.noreply.github.com>