Fix mail move operations after sync
Entity event handling in MailViewModel was written with an expectation
that event batches for moving mails will contain both `Mail` and
`MailSetEntry` updates. However, consequent updates for the same mail
would be optimized away leaving into only the first update event which
does not match this expectation.

We have adjusted MailViewModel in a way that it does not depend on
specific shape of the entity updates. We have introduced a cache for
`MailSetEntry`s so that we can react to entity updates without relying
on other entity events.

Co-authored-by: paw <paw-hub@users.noreply.github.com>