Fix InboxRuleHandler being stuck when called too often
MailViewModel might load and apply inbox rules to the same mails
repeatedly until the cache is updated and InboxRuleHandler wouldn't
apply inbox rules until it's not called for the delay period. This would
lead to a kind of deadlock where inbox rules won't be applied and the
list wouldn't load further.

We replaced `debounce()` with `throttle()` to make sure that the inbox
rules are applied at least every `DEBOUNCE_FIRST_MOVE_MAIL_REQUEST_MS`
which means that attempting to apply inbox rules repeatedly does not
hurt.

Co-authored-by: paw <paw-hub@users.noreply.github.com>