Fix mail view not loading the whole list sometimes
MailViewModel expects that the `EntityClient` will fulfill the contract
of providing the requested amount of items if the list is not loaded
completely yet. However, there are some race conditions where the cache
is updated while we are loading a part of the list from the server so
the end result will not have enough items.

We replaced the check with a check for empty list. This might lead to
additional requests. It should happen though as the cache is already
aware of whether the list is completely loaded.

A more reliable solution could be propagating the information about
whether the list is exhausted explicitly. Another solution could be
locking the range even for individual `put()` operations. Another
solution could be to load part of the list from cache before doing the
network request and not after that.

Co-authored-by: paw <paw-hub@users.noreply.github.com>