Support user area group key rotation for non-admin users
After rotating the user group key for a non-admin users the user also
needs to rotate their user area group keys (mail, contact, calendar..).

The GroupKeyRotationService allows doing that in a way that internal
users can rotate their own area groups if they are non admin users.
Non-admins are not allowed to rotate team or customer groups.

User area groups are not guaranteed to have the adminEncGroupKey set on
the group instance. When doing the user area group key rotation we
always set this value to null now for all user area groups, no matter if
it existed before. We want to achieve that all user area groups have
the same state, which is having no adminEncGroupKey for this group.

tutadb#1844

Co-authored-by: bedhub <bedhub@tutao.de>
Co-authored-by: Vitor Sakaguti <vis@tutao.de>
Co-authored-by: @sashub