Fixes build scripts
After splitting the worker into two, we forgot to set the correct path
for the built worker to worker-bootstrap. Now it gets the correct
location  depending on which app is being built.

It also adds a new test to the github pipeline to ensure that both apps
can be built using the rollup and respect the import rules.