Fixes text cutoff in external content dropdown
In some languages like German or Russian the external content dropdown
text doesn't fit and so it gets cutoff with ellipsis.

This commit fixes that by shifting the dropdown as needed so the entire
text is displayed.
When fitting the entire text on screen isn't possible, the dropdown
spans the whole width and text is cutoff.

Fixes: #7458

Co-authored-by: paw <paw-hub@users.noreply.github.com>