Fix login after clicking on email notification
After rebase of restructuring for app split we were left with two
versions of `openMailbox()`. The one that supports opening mails was
not used anymore.

We swapped the implementation for a real one, moved it into the mail
part of the source tree and refactored both `openMailbox()` and
`openCalendar()` to be classes with injected dependencies.

Fix #7650

Co-authored-by: hrb-hub <181954414+hrb-hub@users.noreply.github.com>