Bypass the cache when getting keys via adminEncGKey
When getting the current group key for a given group, it can be the case
that the group is loaded from the cache before the cache is updated by
the EventBusClient with a new key, so the key that we return is not
actually the current one. Bypassing the cache solves this problem.

It does come with the disadvantage that we may be doing unnecessary
network requests.

tutadb#1883