Make default email signature clickable
The default email signature should have a clickable link so users don't have to copy and paste the link into the browser.
The link itself is added ass a parameter into the translation text. As we don't detect links in email content and make
them clickable automatically, I decided to surround the replacement with an anchor tag.
I also introduced a new link, that was requested by marketing.