Revert "Enable Argon2id migration for all"
This reverts commit 2f550b351224c80e782b25346dc5bc3525be706c.

There is a bug when a different client performs the migration. Existing clients with persistent sessions will use the
wrong key and won't be able to decrypt or authenticate. We need to detect that there was a KDF update and discard the
saved password key in that case. And until all supported clients have this fix we cannot enable the migration.

tutadb#1895