Handle KDF migration from different client
If a different cliend executed a KDF migraiton, the passphrase key
changes and we cannot rely on the stored password key from the
credentials. To successfully resume the session we need to derive the
passphrase key with the updated KDF. We still need to update the
passphrase key in the credentials to allow the SDK to use it (for
decrypting rich notifications).

Notice that between the migration and login in, the rich notifications
cannot be decrypted. We accept this as a minor inconvenience.

tutadb#1895