Check if mail moved after loading cached
First we check if we can re-load the Mail.

Second, we check if the mail's sets no longer contain the same folder as
the cached mail, as the mail will still disappear from view even if the
mail still exists with that ID.

This fixes a regression caused by a previous commit.

Fixes #7897