Allow basic_text for safeStorage backend
On some systems, a backend may not be detected correctly. In a lot of
cases, this can occur if the user is not using one of a hardcoded list
of desktop environments (you can see this list in xdg_util.cc as well
as key_storage_util_linux.cc in GetDesktopEnvironment and SelectBackend
respectively in Chromium's source code), and it determines this using
environment variables.

Basically, if Chromium is unable to detect your desktop environment this
way, safeStorage will not be available in any way except basic_text.

We don't really have any control over this outside of the
preselectGnomeLibsecret() method which may or may not work, but users
can override it with CLI args or just set an app password.

As such, this commit allows basic_text to be used as a backend.

Closes #7257