Fix dropdown not scrolling when there's not enough vertical space
The inner dropdown container was overflowing the outer one instead of
scrolling because its height was not restricted. This could be
reproduced on a screen that's not tall enough e.g. small mobile
screen or landscape orientation of the phone.

Now we set the height to the one of the parent to always have correct
height.

Close #7937