fix opening mail preview from notification on mobile after key rotation, #tutadb1913
We did not set the owner key version properly on an instance when decrypting it. This was fine for owner key version 0 which is the default fallback but it does not work after a key rotation. We now set this properly, so that we can preview the mail without loading it from the server again when opening the app by tapping the notification.