Enable argon2 migration
We enable the argon2 migration for every user that is still using bycrypt
to derive the password key. During the login process we derive a new password key
from the existing password and encrypt the user group key with it.
This step is a requirement for upgrading to quantum safe key material.