Fix concurrency issues with MailViewModel
Add a marker so that we can abort view changes early if state changes,
such as if a different mail/folder view was requested while the current
one was loading.

This can happen if you change folders during entity updates. You can end
up having both changes fight over which view, and this can result in
repeated view switching, as a previous call will end up overriding a
later call.

There are a few ways to improve this. Mailbox model, for example, could
just be fully loaded, and then much of it wouldn't have to be async...

Closes #8031

Co-authored-by: charlag <charlag@users.noreply.github.com>
Co-authored-by: BijinDev <BijinDev@users.noreply.github.com>