fix typeRef for aggregate type defaults to parent app
When serializing entities in json_serializer.rs, we currently look for
a certain typeRef for an aggregate only on the parent app. However, if
the aggregate typeRef is indeed defined in another app (e.g. SystemModel),
we can not find the typeRef and the entity is serialized incorrectly.
This commit fixes this issue, by loading the typeRef from the model in
which the aggregate typeRef is defined.

Co-authored-by: Kinan <104761667+kibibytium@users.noreply.github.com>