add post multiple to blob_facade and StorageModelV10
This commit adds a post multiple to the blob_facade,
in order to post multiple files (e.g. attachments) in a
minimum amount of request to the BlobService. The file data is
serialized inside the binary_blob_wrapper_serializer to a concatenated
binary format. All chunking logic is contained within
the blob_facade and binary_blob_wrapper_serializer.

Co-authored-by: Kinan <104761667+kibibytium@users.noreply.github.com>