Extract new component `MonospaceTextDisplay` from `RecoverCodeField`
This is the "pure" field for monospaced text. It features the optional
attributes `placeholder` and `chunkSize`. `placeholder` will be shown if
the `text` attribute is empty. `chunkSize` splits `text` into chunks
before displaying it.

Theoretically, it would make sense to also use `chunkSize` inside
`RecoverCodeField`, but the copy-to-clipboard feature depends on chunked
text which is impossible to get out of `MonospaceTextDisplay`. So we
keep explicit chunking for now, even if it results in minor code
duplication.