Fix derivation of targetUserGroupKeyAuthKey
This key serves to authenticate the new admin group key to other admins,
and should therefore be accessible by each admin. We have two options
for this: the old admin group key, or each admin's existing user group
key.

The previous implementation derived the authentication key from the
admin group key, which is not what was defined in the design of this
key rotation process.

tutadb#1906