Refactor element-specific behavior out of ListModel
We do not want ListModel to specifically only work on ListElement types,
or even element types in general. As such, we generalize it to a list of
some kind of elements with some kind of ID, even if these are not types
that would be stored remotely as part of the model.

Also renames entity/element to item for the ListModel.

Closes #8222

Co-authored-by: hrb-hub <181954414+hrb-hub@users.noreply.github.com>