[node-mimimi] fix global mutable static warnings
Keep a Console always in existence, but make it so the sender can be
set/unset. We can do this by using Once to ensure that our panic hook
and logger are set once, and then having Console be pre-initialized
with the sender set to None.

This ensures we do not need any static mutables, thus Console can be
inited and deinited safely.

cherry-picked from 2184238aa166b3f739ead9f0529935171d17fe60

Co-authored-by: hrb <hrb-hub@users.noreply.github.com>
Co-authored-by: ivk <charlag@users.noreply.github.com>
Co-authored-by: paw <paw-hub@users.noreply.github.com>