[mail-import] rewrite state handling for mail file import, improve UI
we're moving to persisting the import state only on the server,
with the files to be imported on disk the only local state that
is kept by the importer.

Most errors happen during import preparation after we receive the files
to import. These can be caught & handled immediately. If no errors
happen, the remote state will be set up on the server at this stage and
can be subscribed to by the UI by loading it via its ID

Some other errors can happen asynchronously while the importer is
running in the background. For those, the importer sets a hook to
receive and handle them as they happen.

any normal state updates are sent to the server in node-mimimi and
received in the UI via entity updates.

Co-authored-by: nig <nig@tutao.de>
Co-authored-by: sug <sug@tutao.de>
Co-authored-by: map <mpfau@users.noreply.github.com>
Co-authored-by: jhm <17314077+jomapp@users.noreply.github.com>
Co-authored-by: Kinan <104761667+kibibytium@users.noreply.github.com>
Co-authored-by: abp <abp@tutao.de>