Fix autocomplete search field with user email on chrome
On chrome, clicking the "New email" button causes the search input to
be autofilled with the user's email. This happens even with
autocomplete set to "off", and with the button having a type of "button"

Wrapping the search input in a form isolates it from unrelated buttons
and fixes the issue.

Close #8236