Fix unexpected error during signup for paid account
After subscription to a paid account and using PayPal, the SignupWizard would logout the user before showing the CongratulationPage which cause some backgorund actions to fail and raise exceptions (no user found, no user group key required for initialization of the db).
This commit removes this unecessary logout and resolve a bunch of error related to this race condition.

Co-authored-by: bedhub

tutadb#1805