refactor LanguageViewModel to use new translation types
The reason for this refactor is to make it possible to put "data-testid"
attributes on HTML tags. Those attributes can be used with UI test
frameworks to make automated testing independent of the translation
language or text.

The idea is to get rid of the lazy<string> types and instead use a new type
called Translation wherever possible. Translation replaces the old
TranslationText and contains both translation key and translation text.

The type MaybeTranslation is a legacy type left to make the transition feasible.
It can be replaced with either TranslationKey or Translation in the future.

Co-authored-by: map <mpfau@users.noreply.github.com>