Refactor key derivations
This affects only Key Rotation related derivations, i.e., authentication
keys and the distribution encryption key. We do not change the user
group distribution key derivation, that is a separate change.

This moves all the binding data to the authentication key derivation
step. It doesn't really matter if it is there or in the tagged data, as
long at it is somewhere, but by using it in the key derivation salt we
increase the security of the resulting authentication key.

We also bind the entire MAC tag system together. The idea here is to
make it easy to find the corresponding tag creation and verification in
the code, and also ensure that the parameters passed in are the same.

This also makes it easy to see how key authentication works in general.

We respect the constraint that no classes or functions may be passed in
or out of the facade (due to IPC).

tutadb#1940

Co-authored-by: vis <vis@tutao.de>
Co-authored-by: hec <hec@tutao.de>
Co-authored-by: mab <mab@tutao.de>