WIP Bundle the entire MAC tag system
The idea here is to make it easy to find the corresponding tag creation
and verification in the code, and also ensure that the parameters passed
in are the same.

It should also make it easy to see how key authentication works in
general.

We should respect the constraint that no classes or functions may be
passed in or out of the facade (due to IPC).

The most advanced stage of the idea can be seen on "USER_GROUP_KEY_TAG".
We can adapt it to the others, as well, but further simplifications,
generalizations, and restrictions are certainly possible.

tutadb#1940