Handle keychain errors when removing credentials
This commit adds minimal handling for the case when the device key is
corrupted in some way, and we try to remove the local SSE data.

This does not address the issue that the local data will still be
unusable, this needs to be addressed in a more general way separately.

Close #8280