WIP Bundle the entire MAC tag system
TODO: fix tests

The idea here is to make it easy to find the corresponding tag creation
and verification in the code, and also ensure that the parameters passed
in are the same.

It should also make it easy to see how key authentication works in
general.

We should respect the constraint that no classes or functions may be
passed in or out of the facade (due to IPC).

tutadb#1940