Fix email disappearing after opening from notification
When the list state changes we attempt to find the currently displayed
mail in the list to create a conversation viewer. In case of displaying
email from notification ("sticky mail") the first few list states will
not include the new email. Previously we would remove the conversation
viewer when this happened.

We added a special case to keep displaying the sticky mail, even if it
is not in the list.

Close #8313

Co-authored-by: ivk <ivk@tutao.de>