fix ImportMailState CREATE event not updating the search indexer
After importing mails on the desktop client, those mails were not
indexed properly due to missing of the CREATE entity event. This is due
to the optimization of EntityEvents (CREATE + UPDATE = CREATE).
Previously, the mails were only indexed on ImportMailState UPDATE events
in the MailIndexer, This led to a corrupted state where the imported
mails are never indexed properly and required the deletion of the
offline cache. This fix ensures that the mails are indexed properly.

Co-authored-by: map <mpfau@users.noreply.github.com>