synchronously delete outdated external calendar events
we're seeing some external calendars that change UIDs every day, so it's
causing a lot of request on login. the fact that it's done by the
calendarModel without any concurrency control also leads to users
getting IP banned; the client fires all the delete requests in parallel
which gives the connection pooling no chance to kick in because by the
time the first TLS handshake is done, all requests are in flight already

awaiting each delete request makes the whole process take more time and
more likely to be interrupted. we will have to rethink external
calendar sync to improve this.