Fix being unable to log in after a partially failed login
In some cases LoginFacade does succeed with the partial login, but
some login actions fail on the page side, e.g. loading of
`TutanotaProperties`. This leads to an inconsistent state where the
worker is in the logged in state while the page context is not.

We solve the issue by catching the errors during page context init and
resetting the state on both sides.

In the future we could make the process more robust by loading all the
necessary parts in `LoginFacade` or making some parts lazy-loaded.

Fix #5537