Fix offline login
Offline login was broken because RootInstance for TutanotaProperties is
not cached anymore. This happened in 6acc5ff because of confusion about
semantics of isCachedType() which was meant to be used for range
requests only. For non-range requests we should cache all types, except
the explicitly ignored ones.

We fixed the issue by only checking for custom id in range requests.

Close #8396

Co-authored-by: bir <bir@tutao.de>