[mimimi] improve error state for import UI
the progress estimation for failed imports was not stopped.
this gave the impression that the import was still ongoing.

we now reset the state if an error happens during import preparation
and will pause the running import if an async error happens.
this will allow the user to try to resume it and see the snackbar or
notification if the error happens again.

also adds some local state cleanup if the import fails before being
set up on the server