Deprecate RSA encryption for mails
We still use the keys, but we show a warning if we detect that TutaCrypt
keys should have been used instead.

In practice, this means that if we have TutaCrypt keys at all, the
sender should have used those. However, in order to prevent unnecessary
warnings when doing a key rotation, we consider it normal to receive RSA
encrypted mails in the same session where we did the key rotation.

There is a corner case when we show a warning even though the RSA
encryption was legitimate, and that is if we decrypt an old RSA mail
after rotating to TutaCrypt keys. We don't expect this to occur very
often.

Co-authored-by: hec <hec@tutao.de>
Co-authored-by: mab <mab@tutao.de>

tutadb#1932