update liboqs to version 0.12.0 and use ml-kem instead of kyber
kyber will be dropped with the next release

we need to apply our hack to turn ml-kem into kyber for backward compatibility. this requires some extra kdf and hash step to bind the key to the ciphertext.

we do not want to pull additional dependencies in typescript and liboqs already comes with shake and sha3, so we just do this in C.

we apply a patch to remove fprintf calls from the compiledd liboqs code to not depend on stdio for wasm build

#tutadb1956