[Android] Adds dexmaker to mock classes during Instrumented Tests
To have the SDK we must move the tests from unit to Instrumented Tests
as unit tests aren't able to load our compiled SDK since the target
mismatches (Unit tests require linux-x86-64).

Mockito isn't able to mock final classes and instead of making all of
them open, we can use dexmaker to proxy the classes and be able to
fully mock final classes.

This commit introduces dexmaker as a dependency and also overrides
dexmaker mockito-core with a more updated one.