fix OfflineStorage not encoding custom elementIds correctly as base64Ext
Element and list element instances with customIds must be stored with
their elementIds encoded as base64Ext in the offline database, to
ensure that these elementIds are sortable when running SQL queries.
When received from our server customId are encoded with base64Url,
which is not sortable!
When introducing MailSetEntries, which are using
custom elementIds (customId type), we did refactor the OfflineStorage
in such a way that all publicly exposed methods guarantee the correct
encoding before reading / writing from / to the database. With a later
commit f99bd69ff01e896f3698ae873310c8750cd0d91a some methods in the
OfflineStorage got refactored which led to wrong encodings in certain
scenarios. This commit ensures that custom elementIds are encoded
correctly and are sortable in the OfflineStorage again.