Add wasm files to service worker cache and enforce service worker updates.
After upgrading the liboqs we noticed that the wasm files were not part of the ServiceWorker cache.
This caused an error during the rollout of the new webclient were old webclients were using the new liboqs.wasm
which was not compatible to their client version.

With this commit we add all wasm files to ServicerWorker cache so clients will use all their required files.
To prevent unexpected errors for on old clients, we enforce the update of a new client version when upgrading
from a broken version to a fixed version. For all future client version updates we want to keep the mechanism
that users can do the update later.