[Android] Fix database busy error
This issue usually occurs when the user locks the application while
the app running in foreground and when receiving an email, opening the
application leads to database.lock error. More precisely, writes to db
take longer while the notification process opens the database. In
testing this always happened with `putLastUpdateTime()`.

Resolved this by increasing the busy_timeout in sqlcipher.

close #8536

Co-authored-by: ivk <ivk@tutao.de>