fix mails disappearing from list view, on ImportMailState update
We were re-initializing the MailListModel for the currently
selected folder when receiving an ImportMailState CREATE or
UPDATE entity event. However, this lead to mails disappearing
from the list view of the currently selected folder. We should
only re-initialize the MailListModel in case selected folder
is equal to the import targetFolder from the ImportMaiLState.

Co-authored-by: abp <abp@tutao.de>
Co-authored-by: nig <nig@tutao.de>