Feature flag to include/exclude tests that requires local http server
Some tests in rust sdk and node-mimi requires actual server running
in localhost. We won't have local server running always
( eg: in non-dev machine and also currently in CI )

Include a feature flag that should include such tests as well. if this
feature flag is not present, we mark test as ignored with a message.