Attachments not removed from viewer in drafts
When we remove that last attachment from draft and close the editor it
is only updating on the mail list not on the viewer. This case happens
when there is only one attachment in draft and we try to remove that.
Resolved this by setting the attachments array as empty when the
loadedAttachments size for mail is empty.

close: #5411