Change attachment option from download to open in Mail Editor.
Currently, giving options to download and remove the attachment when the
user is in Mail Editor. Changed that to open and remove. Instead of
downloading attachment user can open the attachment and if the user
still wants to download it is possible to save attachment from  the
viewer mode.

close: #4371