Fix duplicate upgrade dialogue to paid plan when in search view
When hitting Return key after extending search range while on free plan,
the upgrade dialogue is shown twice.
This happens because input is handled in both the input field's
`onkeydown` and in another event listener on the document that is meant
to handle input when focus is shifted outside the DatePicker (like when
clicking away for example). This is a problem because when the first
dialogue appears it grabs focus and since it's outside the DatePicker,
it triggers an additional handling of input.

To fix this, we add the event listeners to the main-view instead of the
entire document, excluding the modal and overlay layers.

Close #8755
Co-authored-by: hrb-hub <181954414+hrb-hub@users.noreply.github.com>