[Android] Fix opening email notification opens latest mail for alias
Opening an email from a notification list always opens the mail for the
most recent notification instead of the selected one.
This happens because we were using the `abs(1 + address.hashCode())`
as the request code for the open mail PendingIntent, which will result
in having the same request code for all email notifications per address

Fixed by generating a unique request code for every intentOpenMail.

Close #8333

Co-authored-by: hrb-hub <hrb-hub@users.noreply.github.com>
Co-authored-by: ivk <ivk@tutao.de>