Remove two FIXMEs
The first fixme says we need to escape special characters from fts5
syntax. However, simply surrounding each token in quotes is sufficient
to do this.

See section 3.1 "FTS5 Strings" here: https://www.sqlite.org/fts5.html
which states that a string may be specified by surrounding it in quotes,
and that special string requirements only exist for strings that are
not in quotes.

We forgot to remove the second fixme, as we had already ordered results
by time.