Remove limit from OfflineStorageSearchFacade
maxResults was intended to limit IndexedDbSearchFacade, and then the
getMoreSearchResults would be called if more results were needed. If
there are no more results, the results would be re-used.

Since we don't have getMoreSearchResults on OfflineStorageSearchFacade,
the incomplete result gets used. We can simply return everything upfront
instead, though, when we do the initial search.

Co-authored-by: bir <bir@tutao.de>
Co-authored-by: hrb <hrb-hub@users.noreply.github.com>