Ensure MailOfflineCleaner only cleans Trash and Spam once old
Previously MailOfflineCleaner would clean Trash and Spam folders
everytime. This was done under the assumption that Trash and Spam were
excluded from search, which wasn't the case, meaning that they had to be
re-indexed and re-retrieved on every login.

Since we decided to keep Trash and Spam included in search, Trash and
Spam are now cleaned same as all other folders.

Close #8837

Co-authored-by: paw <paw-hub@users.noreply.github.com>